wsl_updater.sh: Add a TUI if dialog package is present
Adds a framework for displaying dialog boxes with the dialog package.
Initial implementation in whiptail was dropped due to limitations and was subsequently replaced with dialog.
Could be extended to be as verbose as the normal shell mode. Current implementation looks pretty adequate however

TODO: Add Zenity and Whiptail modes

Signed-off-by: taalojarvi <sreedevan05@gmail.com>